docs(tabline): clarify the structure of `config.format` output
With any types of "control" 'statusline' like sequences (`%#HLname#`,
etc.), computing label substring (used to show partially visible tabs)
becomes not so simple. There doesn't seem to be a reliable, simple-ish,
and *fast* way to compute those.

The only proper solution seems to be to change behavior to not show
partial tabs, but that adds valuable user feedback (especially with wide
tabs present) so best be preserved.

Unfortunately, it means that the tabs can not have any custom
highlighting (like colored icons). This is shame, but together with
other inconveniences their implementation requires, it looks like a good
compromise. The inconveniences are:
- Colored icons requires a lot of custom highlight groups.
  Foreground-only highlight groups in tabline, sadly, not combine with
  background-present groups set via `%#HLname#`. So at least with
  'mini.icons' it would require 9 hues times 2/3/6 tab states
  (current/visible/hidden * unmodified/modified) highlight groups.
- This would require changes to `config.format` signature. It would need
  to know about tab's highlighting in order to reset it after custom
  icon highlighting. So to enable something like
  `%#IconAzureCurrent#󰢱%#MiniTablineCurrent# init.lua` as output.

Resolve #1279